backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
numpy
torch